+++ examples/calculator_features.py
@@ -1,6 +1,6 @@
 from bakery import assert_equal
 from dataclasses import dataclass
-from drafter import route, start_server, Page, Textbox, SubmitButton, LineBreak
+from drafter import route, start_server, Page, TextBox, SubmitButton, LineBreak
 
 
 @dataclass
@@ -14,9 +14,9 @@ class State:
 def index_page(state: State) -> Page:
     return Page(state, [
         "What is the first number?",
-        Textbox("first", "number", state.first_number),
+        TextBox("first", state.first_number, "number"),
         "What is the second number?",
-        Textbox("second", "number", state.second_number),
+        TextBox("second", state.second_number, "number"),
         LineBreak(),
         SubmitButton("Add", add_page),
         SubmitButton("Subtract", subtract_page),
@@ -46,10 +46,10 @@ def subtract_page(state: State, first: str, second: str) -> Page:
 
 assert_equal(index_page(State(0, 0, "")), Page(State(0, 0, ""), [
     "What is the first number?",
-    Textbox("first", "number"),
+    TextBox("first", "", "number"),
     "What is the second number?",
-    Textbox("second", "number"),
-    SubmitButton("Add", "add_page"),
+    TextBox("second", "", "number"),
+    SubmitButton("Add", "", "add_page"),
     LineBreak(),
     "The result is",
     ""
@@ -57,9 +57,9 @@ assert_equal(index_page(State(0, 0, "")), Page(State(0, 0, ""), [
 
 assert_equal(add_page(State(0, 0, ""), "5", "3"), Page(State(5, 3, "8"), [
     "What is the first number?",
-    Textbox("first", "number", 5),
+    TextBox("first", "5", "number"),
     "What is the second number?",
-    Textbox("second", "number", 3),
+    TextBox("second", "3", "number"),
     SubmitButton("Add", "add_page"),
     LineBreak(),
     "The result is",