+++ change_log.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.0] - 2023-09-12
+
+### Added
+
+- New components: TextBox, SelectBox, CheckBox, TextArea
+- Debug data at bottom of page
+- Ability to restore state from links
+
 ## [0.0.2] - 2023-08-30
 
 ### Fixed